Make `WithTimeout` public
