+++ PackageVersions.props
@@ -6,7 +6,7 @@
     <MicrosoftCodeAnalysisCSharpVersion>4.9.2</MicrosoftCodeAnalysisCSharpVersion>
     <MoqVersion>4.18.4</MoqVersion>
     <GdUnitAnalyzersVersion>1.0.0-rc5</GdUnitAnalyzersVersion>
-    <GdUnitAPIVersion>4.4.0-rc9</GdUnitAPIVersion>
+    <GdUnitAPIVersion>4.4.0-rc10</GdUnitAPIVersion>
     <GdUnitTestAdapterVersion>2.1.0-rc3</GdUnitTestAdapterVersion>
   </PropertyGroup>
 </Project>
